Minor: Move test code from `context.rs` into `sql_integration`  (#2473)
