+++ datafusion/core/tests/sql/select.rs
@@ -1188,3 +1188,25 @@ async fn boolean_literal() -> Result<()> {
 
     Ok(())
 }
+
+#[tokio::test]
+async fn unprojected_filter() {
+    let ctx = SessionContext::new();
+    let df = ctx.read_table(table_with_sequence(1, 3).unwrap()).unwrap();
+
+    let df = df
+        .select(vec![col("i") + col("i")])
+        .unwrap()
+        .filter(col("i").gt(lit(2)))
+        .unwrap();
+    let results = df.collect().await.unwrap();
+
+    let expected = vec![
+        "+--------------------------+",
+        "| ?table?.i Plus ?table?.i |",
+        "+--------------------------+",
+        "| 6                        |",
+        "+--------------------------+",
+    ];
+    assert_batches_sorted_eq!(expected, &results);
+}
