+++ datafusion/core/tests/sql/union.rs
@@ -79,3 +79,47 @@ async fn union_all_with_aggregate() -> Result<()> {
     assert_batches_eq!(expected, &actual);
     Ok(())
 }
+
+#[tokio::test]
+async fn union_schemas() -> Result<()> {
+    let ctx =
+        SessionContext::with_config(SessionConfig::new().with_information_schema(true));
+
+    let result = ctx
+        .sql("SELECT 1 A UNION ALL SELECT 2")
+        .await
+        .unwrap()
+        .collect()
+        .await
+        .unwrap();
+
+    #[rustfmt::skip]
+    let expected = vec![
+        "+---+",
+        "| a |",
+        "+---+",
+        "| 1 |",
+        "| 2 |",
+        "+---+"
+    ];
+    assert_batches_eq!(expected, &result);
+
+    let result = ctx
+        .sql("SELECT 1 UNION SELECT 2")
+        .await
+        .unwrap()
+        .collect()
+        .await
+        .unwrap();
+
+    let expected = vec![
+        "+----------+",
+        "| Int64(1) |",
+        "+----------+",
+        "| 1        |",
+        "| 2        |",
+        "+----------+",
+    ];
+    assert_batches_eq!(expected, &result);
+    Ok(())
+}
